ブログ単体取得API

+++ src/handler/resource/pagination_resource.go
@@ -2,7 +2,7 @@ package resource
 
 import "momonga_blog/api"
 
-func MapPaginationToAPI(total int, page int, limit int) api.Pagenation {
+func MapPaginationToAPI(total int64, page int, limit int) api.Pagenation {
 	return api.Pagenation{
 		Total: total,
 		Page:  page,
